refactor: add ftp_impl.rs
