+++ crates/daemon/src/controls/file_transfer/mod.rs
@@ -1,5 +1,6 @@
 mod file_transfer_client;
 mod file_transfer_service;
+mod ftp_impl;
 
 pub use file_transfer_client::FileTransferClientControl;
 pub use file_transfer_service::FileTransferServiceControl;
